# openapi/ogma-api.yaml
openapi: 3.1.0
info:
  title: Vargate Ogma API
  version: "0.1.0"
  description: |
    Public HTTP contract for Ogma (the Vargate Telemetry product).

    This file is the contract source of truth — the backend implements
    against it, the frontend generates TypeScript types from it. See
    `openapi/README.md` for the contract-first change order.

    The shape covers the T4 onboarding flow:
      1. SSO sign-in (Google / Microsoft callback)
      2. Validate the supplied Anthropic admin key
      3. Pick a region
      4. Start the 90-day backfill
      5. Poll backfill status until done

    Plus `/me` for "who am I currently signed in as."

    Subsequent sprints (T5+) extend this file additively — adding new
    paths, new schemas, new tags. Breaking changes (renames, removed
    fields) require a major version bump and a deprecation window.
  contact:
    name: Vargate engineering
    email: engineering@vargate.ai
  license:
    name: Business Source License 1.1
    url: https://github.com/zeroco84/vargate-telemetry/blob/main/LICENSE
servers:
  - url: /api
    description: Same-origin (nginx routes /api/ to the FastAPI gateway).
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: ogma_session
      description: |
        HttpOnly Secure cookie set by `/api/auth/sso/*/callback` on
        successful sign-in. Carries the short-lived JWT. Default
        auth scheme for browser callers.
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        `Authorization: Bearer <JWT>`. Programmatic callers (CI
        scripts, integration tests) that can't use cookies. The JWT
        format is identical to the cookie payload.

  # ─────────────────────────────────────────────────────────────────────────
  # Reusable schemas
  # ─────────────────────────────────────────────────────────────────────────
  schemas:
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Stable machine-readable code (snake_case). Clients branch on this, not on `message`.
          examples: ["validation_error", "invalid_admin_key", "tenant_not_found", "unauthenticated"]
        message:
          type: string
          description: Human-readable, English. Display verbatim or use for diagnostics.
          examples: ["Anthropic rejected the supplied admin key as invalid."]
        details:
          type: array
          description: Per-field validation breakdown when applicable. Empty / omitted for non-validation errors.
          items:
            type: object
            required: [field, issue]
            properties:
              field:
                type: string
                examples: ["admin_key"]
              issue:
                type: string
                examples: ["must start with 'sk-ant-admin01-'"]
      examples:
        - code: validation_error
          message: One or more fields failed validation.
          details:
            - field: admin_key
              issue: "must start with 'sk-ant-admin01-'"
    SsoCallbackRequest:
      type: object
      required: [code, state]
      description: OAuth 2.0 authorization-code callback parameters.
      properties:
        code:
          type: string
          description: The authorization code issued by the provider.
        state:
          type: string
          description: |
            The CSRF token the frontend set before redirecting to the provider.
            Backend verifies it matches the value stored in the short-lived
            state cookie.
    SsoCallbackResponse:
      type: object
      required: [user_id, email]
      description: |
        Returned alongside the `Set-Cookie: ogma_session=...` response
        header. The session cookie is the durable artifact; this body
        is for clients that want to render "signed in as ..." without
        a follow-up `/me` round-trip.
      properties:
        user_id:
          type: string
          description: Stable internal id, NOT the SSO provider's `sub`.
          examples: ["usr_01HXYZ..."]
        email:
          type: string
          format: email
          examples: ["alice@example.com"]
        name:
          type: string
          description: Provider-supplied display name. May be absent.
...
    Region:
      type: string
      enum: [us, eu]
      description: |
        Data-residency region. Once set on a tenant it MUST NOT change —
        the first anchor is region-bound. Future regions (apac, etc.)
        will be added here.
...
    Tenant:
      type: object
      required: [tenant_id, region]
      properties:
        tenant_id:
          type: string
          description: Stable per-tenant identifier (matches `tenants.tenant_id`).
          examples: ["tnt_01HXYZ..."]
        region:
          $ref: '#/components/schemas/Region'
...
    Me:
      type: object
      required: [user_id, email, tenants]
      properties:
        user_id:
          type: string
        email:
          type: string
          format: email
        name:
          type: string
        sso_provider:
          type: string
          enum: [google, microsoft]
        tenants:
          type: array
          description: |
            All tenants this user has access to. Empty array means the
            user is signed in but hasn't completed onboarding yet
            (between SSO callback and `/onboarding/select-region`).
          items:
            $ref: '#/components/schemas/Tenant'
        role:
          type: string
          enum: [admin, member]
          nullable: true
          description: |
            TM4: the caller's role in their bound tenant ('admin' or
            'member'), or null when not yet bound. Admins may write
            budgets, map identities to users, and change roles; members
            are read + self-service. Advisory for UI gating only — the
            backend enforces the gate independently.

    # ─────────────────────────────────────────────────────────────────────
    # Sessions (T5.5) — first dashboard view of ingested content.
    # A Session is one (date, actor) tuple aggregated across the
    # `code_analytics` and `compliance_activities` streams. Admin API
    # usage records are bucket-grain, not actor-grain, so they don't
    # roll up into Sessions.
    # ─────────────────────────────────────────────────────────────────────
...
  responses:
    BadRequest:
      description: Validation failure or malformed input.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthenticated:
      description: Missing or invalid session.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    InternalError:
      description: Unhandled server error. Body is safe-to-display; details land in server logs.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

# ───────────────────────────────────────────────────────────────────────────
# Default security: every endpoint requires either the cookie or a Bearer
# token. Endpoints that don't (the SSO callbacks) override with `security: []`.
# ───────────────────────────────────────────────────────────────────────────
security:
  - cookieAuth: []
  - bearerAuth: []

# ───────────────────────────────────────────────────────────────────────────
# Paths
# ───────────────────────────────────────────────────────────────────────────
paths:
  /auth/sso/google/callback:
    post:
      tags: [auth]
      summary: Google OAuth 2.0 callback
      description: |
        Handles the redirect back from Google after the user
        authenticates. Verifies `state` against the cookie set
        before redirect, exchanges `code` for an ID token, upserts
        the user row, and issues the session cookie.
      security: []
      operationId: ssoGoogleCallback
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SsoCallbackRequest'
      responses:
        '200':
          description: Sign-in successful. Body carries user identity; cookie carries the session.
          headers:
            Set-Cookie:
              description: '`ogma_session=<JWT>; HttpOnly; Secure; SameSite=Lax`'
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SsoCallbackResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          description: |
            State token mismatch (likely CSRF attempt) or the `code`
            was rejected by Google. Returns an ErrorResponse with
            `code: invalid_state` or `code: oauth_exchange_failed`.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalError'
  /auth/sso/microsoft/callback:
    post:
      tags: [auth]
      summary: Microsoft OAuth 2.0 callback
      description: |
        Microsoft Entra (formerly Azure AD) callback. Same flow as
        Google's callback; separate path because the provider's
        token-exchange endpoint and ID-token claim shape differ.
      security: []
      operationId: ssoMicrosoftCallback
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SsoCallbackRequest'
      responses:
        '200':
          description: Sign-in successful.
          headers:
            Set-Cookie:
              description: '`ogma_session=<JWT>; HttpOnly; Secure; SameSite=Lax`'
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SsoCallbackResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          description: State mismatch or `code` rejected by Microsoft.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /me:
    get:
      tags: [me]
      summary: Return the signed-in user's profile + tenant binding
      description: |
        The frontend uses this on every route entry to confirm the
        session is still valid and to pick up tenant changes (e.g.,
        after `select-region` updates the JWT's `tenant_id` claim).
      operationId: getMe
      responses:
        '200':
          description: User profile + accessible tenants.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Me'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '500':
          $ref: '#/components/responses/InternalError'
...